github-files/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
pymongo==4.6.0
pydantic==2.5.0
python-multipart==0.0.6